+++ backend/src/api/gen_stats.py
@@ -3,8 +3,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, Request
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from back_api.rate_limiter import limiter
-from back_api.schemas import StatisticsResponse
+from api.rate_limiter import limiter
+from api.schemas import StatisticsResponse
 from database.database import get_session
 from service.record import RecordService, get_record_service
