+++ backend/src/api/publ_from_hash.py
@@ -4,8 +4,8 @@ from typing import Annotated
 from fastapi import APIRouter, Depends, HTTPException, Request
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from back_api.rate_limiter import limiter
-from back_api.schemas import PublResponse, RecordHashRequest
+from api.rate_limiter import limiter
+from api.schemas import PublResponse, RecordHashRequest
 from database.database import get_session
 from database.hash import decrypt_id
 from service.record import RecordService, get_record_service
